storage: accept bytes in Volume create() and _create()
Convert Volume.create() and Volume._create() to accept bytes for size in
initial_size.

If the size of volumes not aligned to underlying disk block size, it can
cause problems, therefore we round provided values to 4k block size.

Also added a test which checks that the size is really aligned to 4k
block size.

Change-Id: Ib3d36d083f898c695a78b4efcd9b13d4f44c9d07
Bug-Url: https://bugzilla.redhat.com/1592916
Signed-off-by: Vojtech Juranek <vjuranek@redhat.com>