storage: accept bytes in Volume.newMetadata()
Volume.newMetadata() now accepts values in bytes.

As we replace appropriate call in make_volume in testlib.
Alse a bug in make_block_volume is fixed in this patch. Before volume
size was adjusted for all volume type, while it is relevant only for
RAW volumes. Volume size is now adjusted only for RAW volumes.
This also resulted in fixing one more test where expected optimal size
was incorrect as the chain was created with wrong volume sizes.

Change-Id: I506cfe33937384aecea92d32614cb44695fd6c03
Bug-Url: https://bugzilla.redhat.com/1592916
Signed-off-by: Vojtech Juranek <vjuranek@redhat.com>