tests: run volume tests on 4k storage
This patch intorduces a fixture which provides function for creating
a storage domain on top of user mount to avoid repeating code. Fixture
use user_mount fixture, so that the tests now run on all storage
backends.

As some tests are only for SD version 5, we intorduce also fixtures
user_mount_v5 and user_domain_v5 to avoid skipping unwanted combinations
in the test.

Change-Id: I0eac5de6a2c442f934a3d21cb95386665b01e322
Bug-Url: https://bugzilla.redhat.com/1592916
Signed-off-by: Vojtech Juranek <vjuranek@redhat.com>