storage: decode iscsiadm output
Running iscsiadm command returns output in bytes. Several functions in
iscsiadm module parse this output and deal with it as a string. This
fails in python3 - example error trace back which shows several issues
in this module is bellow.

Decodes command output as UTF-8 string and return this string from the
command function. Also make this function public so we can test it.

The patch also includes simple test for the cmd function which verifies
that returned output is string in python2 as well as in python3.

Examples error output:

      File "/usr/lib/python3.6/site-packages/vdsm/storage/iscsiadm.py", line 205, in iface_info
	if line.startswith("#"):
    TypeError: startswith first arg must be bytes or a tuple of bytes, not str

      File "/usr/lib/python3.6/site-packages/vdsm/storage/iscsiadm.py", line 189, in iface_list
	for value in re.split('[\s,]', line))
      File "/usr/lib64/python3.6/re.py", line 212, in split
	return _compile(pattern, flags).split(string, maxsplit)
    TypeError: cannot use a string pattern on a bytes-like object

Change-Id: Ib8b05b8af86a9bb445699ca6874b4f9703d9772e
Signed-off-by: Vojtech Juranek <vjuranek@redhat.com>