tests: add test for multipath resize_map()
Output of multipath is either `ok` or `fail` on stdout:

    # multipathd resize map 36006048ca9da4a4597bdaae943f78957
    ok

    # multipathd resize map 36006048ca9da4a4597bdaae943f78957222
    fail

The patch also add a general fixture which prepares a shell script
in tmp directory, which can be used by another fixture to fake an
executable. Typical usage is that this fixture will monkeypatch
executable path and replace it by path of this script. Script will
fake the output of the executable.

Change-Id: I4cdff43e40d8c116f8faf39abd77ce4cd191b82e
Signed-off-by: Vojtech Juranek <vjuranek@redhat.com>