vdsmd.service: Ensure that child processes are killed
Vdsm depends on systemd to terminate all child processes when vdsmd
service is stopped. In EL6 time, we deopended on cpopen to set a death
signal for storage child processes, ensuring that commands writing to
storage are killed when vdsm is stopped, even if vdsm is killed.

While moving to EL7 and systemd and porting to python 3, we removed
death signal support, and started to depend on systemd to kill child
processes.

commit b9fd772eafcc7b60f3f0e4965183a9da6ef05c97
Author: Yaniv Bronhaim <ybronhei@redhat.com>
Date:   Sat Jan 23 19:04:22 2016 +0200

    Remove deathSignal usages in sync execCmd calls

Depending on systemd to kill child processes[1] is more correct than
using death signal, since death signal affects only the direct child
processes created by vdsm. If a child process creates its own child
processes, they are not using death signal, and would not be killed when
vdsm is terminated.

Turns out that systemd does not terminate child processes if
ExecStopPost is set[2]. This can leave commands such as qemu-img writing to
to logical volume after vdsm was stopped. Engine will detect that the
command has failed and will delete the logical volume. A new logical
volume may share the same storage as the old logical volume, causing the
left-over qemu-img to write to another volume, possibly corrupting user
data.

I think the most robust way to fix this issue would be to avoid the
dependency on systemd and use volume leases held by the actual program
accessing storage (e.g. qemu-img), and the SPM commands deleting
volumes. This requires major changes in storage code, that can be
planned for future version. For now, we need a minimal change that can
be ported safely to 4.3 and work with any supported engine version.

Fix this issue by removing the ExecStopPost from vdsmd service file.
With this change, systemd terminates child processes as expected.

ExecStopPost was used only for calling after_vdsm_stop hook. Replace it
by calling the hook from vdsm main thread, when starting shutdown after
receiving a termination signal. The downside of this change is that
after_vdsm_stop will not be called if vdsm is killed.

Systemd behavior was tested only on Fedora 29, with the reproducer
scripts (see bug 1761260). This change is not tested yet.

[1] http://man7.org/linux/man-pages/man5/systemd.kill.5.html
[2] https://bugzilla.redhat.com/show_bug.cgi?id=1761260#c0

Change-Id: Icf4e2ecc35065fe93a2489c68c6717724f1e793a
Bug-Url: https://bugzilla.redhat.com/1759388
Signed-off-by: Nir Soffer <nsoffer@redhat.com>