storage: activate/deactivate LVs before/after zeroing
When preparing LVs for removal, we amend their tags and then activate
them. This is not needed if remove them immediately. Remove activation
of LVs from this function and call it when it really needed, before we
zero the LVs. Also deactivate LVs after that, so that we are not
removing active LVs.

Activate LVs one by one instead of bulk activation. If activation of one
of the LVs fails, whole operation fails. When we do it separately for
each LV, only faulty LVs fail. Running single lvm command for each LV
may be little bit worse from perfomance point of view, but reliability
is more important here.

Change-Id: Id609e11163862552bde63f07af5c85957c1dfcbf
Bug-Url: https://bugzilla.redhat.com/1639360
Signed-off-by: Vojtech Juranek <vjuranek@redhat.com>