storage: always call lvm.removeLVs() with tuple of LVs
lmv.removeLVs() accepts iterable of LVs to be removed. It can be called
also with single string value representing single LV as it calls
lvm.normalize_args() before using it. However, this is just a workaround
how to make it iterable and one should always pass only a tuple of LVs
into this function so that normalize_args() can be removed in the
future.

Replace normalize_args() function with a check that provises lvs is
iterable (either list or tuple).

Change-Id: Ibc50e3878263888235499a5a9939e1399de64688
Signed-off-by: Vojtech Juranek <vjuranek@redhat.com>