storage: don't hardcode device-mapper major number
Major number for device-mapper devices is usually 253, but in some rare
cases (e.g. VMs installed via virt-builder), this number is different.
Range 240-254 is for experimental usage and number 253 is not reserved
for device-mapper, see [1] for reserved numbers and more details.

Don't hardcode the number but find it out on every host by reading
/proc/devices and searching for device-mapper major number.

[1] https://www.kernel.org/doc/Documentation/admin-guide/devices.txt

Change-Id: I490e361e1030ae70ca442c645a715e42ee2623db
Signed-off-by: Vojtech Juranek <vjuranek@redhat.com>