tests: add fake prepare and tear down volume methods
Add fake prepareImage() and teardownImage() methods which will be later
on used for testing new flows for changing CDs located on block SD.

Add also test for changing CD on block device. This flow is currently
not used as CD is always specified as path but not as PDIV. This is
now mainly to test newly added fake methods.

Change-Id: Ic02040d2b59836cc156a9b098ed15ac10593d030
Signed-off-by: Vojtech Juranek <vjuranek@redhat.com>