virt: add `change` key into metadata
Add change key into virt metadata. This element will be used during
change of CD in CDROM and it's purpose is to store information about
volume being loaded into CD. This information can be used in recovery
process to (de)activate volumes on block storage. E.g. if vdsm fails
during change CD on block storage once we activate new CD volume but
haven't deactivated old CD volume yet, we end up with two active
volumes and have to deactivate one of them. So far we keep information
only about loaded CD. With this keep information about old as well as
new CD and act accordingly during recovery process. Once new CD is
successfully changed by libvirt and old CD volume is deactivated,
metadata will be updated with information about current CD and `change`
element will be removed.

Current implementation
======================

Currently, change CD is exposed as part of vdsm API and accepts two
parameters from the called - vm ID, begin UUID of the VM and driveSpec,
begin specification of the new CD. driveSpec is a dictionary with
following items: `iface`, `index` and `path`. `iface` and `index`
specifies CDROM device, `path` specifies location (path) of new CD on
the disk as seen by host. Empty path (i.e. `path=""`) means that CD
should be ejected. Example of the the driveSpec for changing CD is as
follows:

    {
        'iface': 'sata',
        'path': '/rhev/data-center/mnt/blockSD/5f2f58dd-4c46-4f38-88d3-d6a78973c0e2/images/2eb0aa33-f5ff-4fbf-a149-da17c4d191d9/28c61447-bcd4-48e8-b6d4-2613d4a50bff',
        'index': '2'
    }

Example of the the driveSpec when CD is ejected is as follows:

    {
        'iface': 'sata',
        'path': '',
        'index': '2'
    }

When the CD is loaded, metadata with PDIV information and volume chain
is stored into VM metadata. Example of metadata:

    <ovirt-vm:device devtype="disk" name="sdc">
        <ovirt-vm:domainID>88252cf6-381e-48f0-8795-a294a32c7149</ovirt-vm:domainID>
        <ovirt-vm:imageID>89f05c7d-b961-4935-993f-514499024515</ovirt-vm:imageID>
        <ovirt-vm:poolID>13345997-b94f-42dd-b8ef-a1392f65cebf</ovirt-vm:poolID>
        <ovirt-vm:volumeID>626a493f-5214-4337-b580-96a1ce702c2a</ovirt-vm:volumeID>
        <ovirt-vm:volumeChain>
            <ovirt-vm:volumeChainNode>
                <ovirt-vm:domainID>88252cf6-381e-48f0-8795-a294a32c7149</ovirt-vm:domainID>
                <ovirt-vm:imageID>89f05c7d-b961-4935-993f-514499024515</ovirt-vm:imageID>
                <ovirt-vm:leaseOffset type="int">105906176</ovirt-vm:leaseOffset>
                <ovirt-vm:leasePath>/dev/88252cf6-381e-48f0-8795-a294a32c7149/leases</ovirt-vm:leasePath>
                <ovirt-vm:path>/rhev/data-center/mnt/blockSD/88252cf6-381e-48f0-8795-a294a32c7149/images/89f05c7d-b961-4935-993f-514499024515/626a493f-5214-4337-b580-96a1ce702c2a</ovirt-vm:path>
                <ovirt-vm:volumeID>626a493f-5214-4337-b580-96a1ce702c2a</ovirt-vm:volumeID>
            </ovirt-vm:volumeChainNode>
        </ovirt-vm:volumeChain>
    </ovirt-vm:device>

Proposed changes
================

To be able to work with images on block SD, we need PDIV of the images.
Therefore it would be useful to request this information directly from
the caller instead of `path`. The API should still require a dict with
`iface` and`index`, but caller should provide a dict, `pdiv` with PDIV
information. To keep backward compatibility, `path` should still be a
valid option. The `DriveSpecCdrom` needs to be extended with attribute
`drive` of a type `DriveSpecVolume` wiht defautl value `null` to keep
backward compatibility.

As `path` is kept only for backward compatibility and won't used by new
implementation, for new calls it will be removed from RPC call dict.

Request to eject CD will be done by not providing PDIV information in
`drive` dict., i.e. by sending `null` for the `drive`.

Example of newly proposed structure of dict is as follows:

    {
        'iface': 'sata',
        'index': '2',
        'drive': {
            'device': 'cdrom',
            'poolID': 13345997-b94f-42dd-b8ef-a1392f65cebf'',
            'domainID': '88252cf6-381e-48f0-8795-a294a32c7149',
            'imageID': '89f05c7d-b961-4935-993f-514499024515',
            'volumeID': '626a493f-5214-4337-b580-96a1ce702c2a'
        }
    }

Example of the dict when eject CD is requested is as follows:

    {
        'iface': 'sata',
        'index': '2',
        'drive': null
    }

Information from the dict will be stored in <change> element of the
metadata and after successful change of CD, the <change> element will
be removed and new PDIV will be stored in CD metadata instead of old
PDIV.

As current implementation of metadata doesn't allow to store empty
elements, we need to store additional element <state> to know what it
currently running operation. Without this, when ejecting CD, <change>
element would be empty and where not stored, and thus in case of failure
we would loose the information that CD eject change was in progress.

As we actually don't need volumeChain element for anything, this
element can be removed. Final CD metadata will look like this when
CD change is in the progress:

    <ovirt-vm:device devtype="cdrom" name="sdc">
        <ovirt-vm:domainID>88252cf6-381e-48f0-8795-a294a32c7149</ovirt-vm:domainID>
        <ovirt-vm:imageID>89f05c7d-b961-4935-993f-514499024515</ovirt-vm:imageID>
        <ovirt-vm:poolID>13345997-b94f-42dd-b8ef-a1392f65cebf</ovirt-vm:poolID>
        <ovirt-vm:volumeID>626a493f-5214-4337-b580-96a1ce702c2a</ovirt-vm:volumeID>
        <ovirt-vm:change>
                <ovirt-vm:state>loading</ovirt-vm:state>
                <ovirt-vm:domainID>09a0c152-7b9a-44ae-866c-c6486d3e10c9</ovirt-vm:domainID>
                <ovirt-vm:imageID>1e1ff19c-01d0-432c-97c2-5883fc833ce4</ovirt-vm:imageID>
                <ovirt-vm:poolID>fb7797bd-3cca-4baa-bcb3-35e8851015f7</ovirt-vm:poolID>
                <ovirt-vm:volumeID>25b29335-8759-4464-a35f-023bcfd922c8</ovirt-vm:volumeID>
        </ovirt-vm:change>
    </ovirt-vm:device>

If the change succeeds, the metadata will be:

    <ovirt-vm:device devtype="cdrom" name="sdc">
        <ovirt-vm:domainID>09a0c152-7b9a-44ae-866c-c6486d3e10c9</ovirt-vm:domainID>
        <ovirt-vm:imageID>1e1ff19c-01d0-432c-97c2-5883fc833ce4</ovirt-vm:imageID>
        <ovirt-vm:poolID>fb7797bd-3cca-4baa-bcb3-35e8851015f7</ovirt-vm:poolID>
        <ovirt-vm:volumeID>25b29335-8759-4464-a35f-023bcfd922c8</ovirt-vm:volumeID>
    </ovirt-vm:device>

When ejecting CD is in progress, the metadata will be:

    <ovirt-vm:device devtype="cdrom" name="sdc">
        <ovirt-vm:domainID>88252cf6-381e-48f0-8795-a294a32c7149</ovirt-vm:domainID>
        <ovirt-vm:imageID>89f05c7d-b961-4935-993f-514499024515</ovirt-vm:imageID>
        <ovirt-vm:poolID>13345997-b94f-42dd-b8ef-a1392f65cebf</ovirt-vm:poolID>
        <ovirt-vm:volumeID>626a493f-5214-4337-b580-96a1ce702c2a</ovirt-vm:volumeID>
        <ovirt-vm:change>
                <ovirt-vm:state>ejecting</ovirt-vm:state>
        </ovirt-vm:change>
    </ovirt-vm:device>

If ejecting of the CD succeeds, there won't be any CD related metadata,
as we haven't anything to store there.

Change-Id: Ia8366d4962aa7f17f52aaf90c2b9fb8881c0aa42
Bug-Url: https://bugzilla.redhat.com/1589763
Signed-off-by: Vojtech Juranek <vjuranek@redhat.com>