vm: add new private change_cd method
Add new implementation of change CD functionality. The method is private
as will be called only for corresponding cluster version and when CD
volume is specified using PDIV. Otherwise old implementation will be
called to keep backward compatibility. This is also the reason to keep
old implementation.

Main difference is that the new method adds CD metadata about loaded
CD. In case of failure we can use this metadata to recover CD and also
properly tear down the unused volumes.

Change-Id: I9166404209484b29f8394b44cbc1b46291a4720a
Bug-Url: https://bugzilla.redhat.com/1589763
Signed-off-by: Vojtech Juranek <vjuranek@redhat.com>