automation: don't use unsafe pushd
Currently CI fails with

    OSError: [Errno 16] Device or resource busy: '/var/lib/mock/epel-8-x86_64-6fbd7704c310804ff7016833dcc4221d-3162111/root/var/tmp/vdsm-storage/mount.file-512'

which is a result of failed userstorage teardown, which fails with

    python3: can't open file 'tests/storage/userstorage.py': [Errno 2] No such file or directory

which is related to combined coverage command failure:

    Coverage.py warning: Couldn't read data from '/home/jenkins/workspace/vdsm_standard-check-patch/vdsm/tests/.coverage-gluster': UnicodeDecodeError: 'utf-8' codec can't decode byte 0x96 in position 106: invalid start byte

The automation script changes the directory to `tests` before running
combined coverage command and after it moves back to original dir.
However, if the command fails, it doesn't move back and following
storage teardown fails as well as the script is in wrong directory and
cannot find userstorage script.

Don't use unsafe `pushd` command, but run coverage commands in subshell.
Running `cd` in a subshell doesn't impact parent shell and therefore
even in case of failure, tear down should work as expected.

Change-Id: I02ff1324b1834ab88c230929422ef88a20e4807f
Signed-off-by: Vojtech Juranek <vjuranek@redhat.com>