automation: let combined coverage command always succeed
In previous patch execution of teardown was fixed, but CI still fails,
as failing coverage fails whole build:

    22:45:21 Coverage.py warning: Couldn't read data from '/home/jenkins/workspace/vdsm_standard-check-patch/vdsm/tests/.coverage-virt': UnicodeDecodeError: 'utf-8' codec can't decode byte 0x96 in position 106: invalid start byte
    22:45:21 + teardown
    22:45:21 + res=1
    22:45:21 + '[' 1 -ne 0 ']'
    22:45:21 + echo '*** err: 1'
    22:45:21 *** err: 1

Make coverage commands always succeed as if the tests pass, job should
succeed. Coverage error is clearly visible in the logs and one should
note it when checking summary coverage report and there's non (and if
nobody checks coverage report, we probably don't need it:-).

Change-Id: If0a5e3daa2cb160d1eceb7d0868af422c02016c3
Signed-off-by: Vojtech Juranek <vjuranek@redhat.com>